refactor: split worker sessions HTTP adapter
